Improvement in the parsing and ingestion process
The JSON parsing and addition process has been improved by switching to a concurrent model, which greatly improves efficiency and speed. Additionally, indexes have been implemented to increase the speed of search operations.

The number of neo4j sessions to use (Workers) and the number of retries in case of deadlock are now configurable through the GUI.